% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_selection.R
\name{backward_selection}
\alias{backward_selection}
\title{Finds a subset of predictors using backward selection}
\usage{
backward_selection(X, y, n_predictors, model = lm, verbose = TRUE)
}
\arguments{
\item{X}{Dataset X}

\item{y}{Dataset y}

\item{model}{Modeling function}

\item{verbose}{Whether to print information of the selection process}

\item{n_predicors}{Amount of predictors to select}
}
\value{
A stepwise_selection object
}
\description{
Finds a subset of predictors using backward selection
}
