% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cv_lambda.R
\name{calculate_cv_lambda}
\alias{calculate_cv_lambda}
\title{Calculates the Lambda with M-fold Cross Validation}
\usage{
calculate_cv_lambda(M, X, Y, lambda, regmodel = ridge)
}
\arguments{
\item{M}{Split X into M-many equally sized Groups.
M = 5 or M = 10 is recommended.}

\item{X}{Dataset X has the class "tibble".}

\item{Y}{Dataset Y has the class "tibble".}

\item{lambda}{Lambda is an atomic vector of double values.
All of them are tested to find the optimal lambda with regards to the chosen
regmodel}

\item{regmodel}{Regressionmodel that is used on X and Y. Options are:
ridge and lasso}
}
\value{
A double value
}
\description{
Calculates the Lambda with M-fold Cross Validation
}
