% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_R2.R
\name{calculate_R2}
\alias{calculate_R2}
\title{Calculate model R-Squared}
\usage{
calculate_R2(y, yhat)
}
\arguments{
\item{y}{Response vector (observed values)}

\item{yhat}{Fitted values (predicted values)}
}
\value{
R-Squared statistic
}
\description{
Calculate model R-Squared
}
\examples{
y <- c(3, 5, 2, 8, 7)
yhat <- c(2.8, 5.1, 2.4, 7.8, 7.2)
calculate_R2(y, yhat)
}
