% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.cv.R
\name{coef.cv}
\alias{coef.cv}
\title{Extract coefficients created during the cross validation}
\usage{
\method{coef}{cv}(object, unscale = FALSE)
}
\arguments{
\item{object}{A cv S3 Object}

\item{unscale}{If TRUE, the coefficients will be scaled back to the original
scale}
}
\value{
ridge or lasso coefficients from regobject, depending on which one
was used for the cross validations.
}
\description{
Extract coefficients created during the cross validation
}
