% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.LAR.R
\name{coef.LAR}
\alias{coef.LAR}
\title{Extract coefficients from an estimated least angle regression model}
\usage{
\method{coef}{LAR}(object)
}
\arguments{
\item{object}{A LAR S3 Object}
}
\value{
LAR coefficients from regobject
}
\description{
Extract coefficients from an estimated least angle regression model
}
