% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.lasso.R
\name{coef.lasso}
\alias{coef.lasso}
\title{Extract coefficients from an estimated least angle regression model}
\usage{
\method{coef}{lasso}(object, unscale = FALSE)
}
\arguments{
\item{object}{A LAR S3 Object}

\item{unscale}{If TRUE, the coefficients will be scaled back to the original
scale}
}
\value{
LAR coefficients from regobject
}
\description{
Extract coefficients from an estimated least angle regression model
}
