% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.ridge.R
\name{coef.ridge}
\alias{coef.ridge}
\title{Extract coefficients from an estimated ridge model}
\usage{
\method{coef}{ridge}(object)
}
\arguments{
\item{object}{A ridge S3 Object}
}
\value{
ridge coefficients from regobject
}
\description{
Extract coefficients from an estimated ridge model
}
