% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.stepwise_selection.R
\name{coef.stepwise_selection}
\alias{coef.stepwise_selection}
\title{Extract coefficients from selection_model}
\usage{
\method{coef}{stepwise_selection}(object)
}
\description{
Extract coefficients from selection_model
}
