% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cv_lambda.R
\name{cross_validation}
\alias{cross_validation}
\title{Calculates the Lambda with M-fold Cross Validation}
\usage{
cross_validation(X, Y, M, lambda, regmodel = ridge)
}
\arguments{
\item{X}{Dataset X has the class "matrix".}

\item{Y}{Target variable Y has the class "matrix".}

\item{M}{Split X into M-many equally sized Groups.
M = 5 or M = 10 is recommended.}

\item{lambda}{Lambda is an atomic vector of double values.
cross_validation will find the optimal lambda with regards to the chosen
regression model.}

\item{regmodel}{Regression model that is used on the data during Cross Validation.
Options are:
\itemize{
\item ridge
\item lasso
}}
}
\value{
A list with the mean squared prediction error, the optimal lambda and
the coefficients of the chosen regression model with the optimal lambda.
}
\description{
Calculates the Lambda with M-fold Cross Validation
}
