% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_selection.R
\name{forward_selection}
\alias{forward_selection}
\title{Finds a subset of predictors using forward selection}
\usage{
forward_selection(X, y, n_predictors, model_fct = lm, verbose = TRUE)
}
\arguments{
\item{X}{Dataset X}

\item{y}{Dataset y}

\item{model_fct}{Modeling function with formula and data parameter (like lm)}

\item{verbose}{Whether to print information of the selection process}

\item{n_predicors}{Amount of predictors to select}
}
\value{
A stepwise_selection object
}
\description{
Finds a subset of predictors using forward selection
}
