% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_dummy.R
\name{is_dummy}
\alias{is_dummy}
\title{Check if a column of a data.frame is a dummmy variable}
\usage{
is_dummy(data)
}
\arguments{
\item{data}{data matrix}
}
\description{
Check if a column of a data.frame is a dummmy variable
}
