% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{lasso}
\alias{lasso}
\title{Calculate the Lasso estimates}
\usage{
lasso(X, y, lambda, tol = 1e-07, verbose = F)
}
\arguments{
\item{X}{data frame containing the covariates}

\item{lambda}{penalty parameter ($\lambda > 0$)}

\item{tol}{tolerance level for convergence (default = 1e-5)}

\item{Y}{response vector}
}
\value{
list ..... TODO
}
\description{
Calculate the Lasso estimates
}
