% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_cv.R
\name{lasso_cv_calculation}
\alias{lasso_cv_calculation}
\title{Wrapper Function for the regmodelsuit package.}
\usage{
lasso_cv_calculation(X, y, tol = 1e-07)
}
\value{
List including minimum lambda
}
\description{
Wrapper Function for the regmodelsuit package.
}
