% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/least_angle_regression.R
\name{least_angle_regression}
\alias{least_angle_regression}
\title{Calculate the LARS path}
\usage{
least_angle_regression(X, y, iter = 10)
}
\arguments{
\item{X}{data frame containing the covariates}

\item{Y}{response vector}
}
\value{
list ..... TODO
}
\description{
Calculate the LARS path
}
