% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lasso.R
\name{plot.lasso}
\alias{plot.lasso}
\title{Plot the Lasso Regression Function over a predictor}
\usage{
\method{plot}{lasso}(lassoobj, predictor)
}
\arguments{
\item{lassoobj}{A lasso object}

\item{predictor}{The index or name of the predictor, which will be on the
x-axis}
}
\description{
Pass in a lasso object and select a predictor. The plot contains the original
points of the model and draws the estimation of y based solely on the selected
predictors. \cr
plot.ridge can plot real polynomials, so a model that was fitted for the formula
\code{y ~ x1 + x1^2 + 0.2*x1^5 + x2} could be plotted along x1.
}
\details{
See \code{vignette("plot", package = "regmodel")} for detailed worked examples.
}
