% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ridge.R
\name{plot.ridge}
\alias{plot.ridge}
\title{Plot the Regression Function over a predictor}
\usage{
\method{plot}{ridge}(ridgeobj, predictor)
}
\arguments{
\item{ridgeobj}{A ridge object}

\item{predictor}{The index or name of the predictor, which will be on the
x-axis}
}
\description{
Plot the Regression Function over a predictor
}
