% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.stepwise_selection.R
\name{plot.stepwise_selection}
\alias{plot.stepwise_selection}
\title{Plots the selection process of a stepwise selection object.}
\usage{
\method{plot}{stepwise_selection}(object)
}
\arguments{
\item{object}{Stepwise selection object}
}
\value{
A ggplot.
}
\description{
Plots the selection process of a stepwise selection object.
}
