% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ridge.R
\name{predict.ridge}
\alias{predict.ridge}
\title{Predicted values based on ridge model object}
\usage{
\method{predict}{ridge}(object, newdata = NULL)
}
\arguments{
\item{object}{S3 Object which inherits "ridge"}

\item{newdata}{An optional data frame in which to look for variables with
which to predict. If omitted, the fitted values are used.}
}
\value{
A vector of predictions
}
\description{
Predicted values based on ridge model object
}
