% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.lasso.R
\name{print.lasso}
\alias{print.lasso}
\title{print function for the lasso object}
\usage{
\method{print}{lasso}(x)
}
\arguments{
\item{x}{a S3 lasso object estimated by regmodel}
}
\description{
print function for the lasso object
}
