% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.lasso_cv.R
\name{print.lasso_cv}
\alias{print.lasso_cv}
\title{Print function for the lasso_cv object}
\usage{
\method{print}{lasso_cv}(x)
}
\arguments{
\item{x}{a S3 lasso_cv object estimated by lasso_cv}
}
\description{
Print function for the lasso_cv object
}
