% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ridge.R
\name{print.ridge}
\alias{print.ridge}
\title{Print function for the ridge object}
\usage{
\method{print}{ridge}(ridgeobj)
}
\arguments{
\item{x}{a S3 ridge object estimated by ridge}
}
\description{
Print function for the ridge object
}
