% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ridge_cv.R
\name{print.ridge_cv}
\alias{print.ridge_cv}
\title{Print function for the ridge_cv object}
\usage{
\method{print}{ridge_cv}(x)
}
\arguments{
\item{x}{a S3 ridge_cv object estimated by ridge_cv}
}
\description{
Print function for the ridge_cv object
}
