% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.stepwise_selection.R
\name{print.stepwise_selection}
\alias{print.stepwise_selection}
\title{Prints information of a stepwise selection object.}
\usage{
\method{print}{stepwise_selection}(object)
}
\arguments{
\item{object}{Stepwise selection object}
}
\description{
Prints information of a stepwise selection object.
}
