% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_data_search.R
\name{recursive_data_search}
\alias{recursive_data_search}
\title{Recursively search parent environments for data}
\usage{
recursive_data_search(data, env)
}
\arguments{
\item{data}{Name of data object}

\item{env}{current environment for recrusive search}
}
\description{
Recursively search parent environments for data
}
