% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmodelsuite-package.R
\docType{package}
\name{regmodelsuite-package}
\alias{regmodelsuite}
\alias{regmodelsuite-package}
\title{regmodelsuite: Estimate regression models}
\description{
Implements common penalty based estimation methods i.e Ridge, Lasso, Forward selection, Backward selection and Least angle regresion
}
\author{
\strong{Maintainer}: Jonas Veit \email{mee322@uni-heidelberg.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
