% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridge}
\alias{ridge}
\title{Ridge regression}
\usage{
ridge(X, y, lambda = 0)
}
\arguments{
\item{X}{A matrix of predictor variables.}

\item{y}{A vector of response values.}

\item{lambda}{A numeric value for the regularization parameter.}
}
\value{
S3 object with class ridge, containing the coefficients and other information.
}
\description{
This function calculates the coefficients of a ridge regression model
for a given predictor matrix \code{X} and response vector \code{y}.
}
\examples{
\preformatted{set.seed(21)
X <- matrix(runif(100 * 10), 100, 10)
y <- X \%*\% runif(10, min = -1, max = 1) + rnorm(100, mean = 0, sd = 0.3)
ridge(X,y,0.1)
}
}
