% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_cv.R
\name{ridge_cv}
\alias{ridge_cv}
\title{m-fold Cross Validation with ridge regression}
\usage{
ridge_cv(X, y, m = 10, lambda = NULL)
}
\arguments{
\item{X}{A matrix of predictor variables.}

\item{y}{A vector of response values.}

\item{m}{An integer. Split X into m-many nearly equally sized groups.
m = 5 or m = 10 is recommended. The default is m = 10.}

\item{lambda}{An atomic vector of double values.
cross_validation will find the optimal lambda among those. If no lambda is chosen,
then a default range of lambdas will be used.}
}
\value{
S3 object with class ridge_cv, containing the optimal lambda and other
information.
}
\description{
This function calculates the lambda for which the mean squared prediction error
of ridge regression is the smallest for  a given predictor matrix \code{X} and
response vector \code{y}.
}
