% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor_selection.R
\name{select}
\alias{select}
\title{Finds a subset of predictors using forward or backward selection}
\usage{
select(X, y, n_predictors, model = lm, direction = "forward", verbose = TRUE)
}
\arguments{
\item{X}{Dataset X}

\item{model}{Modeling function}

\item{direction}{Selection direction, either "forward" or "backward"}

\item{verbose}{Whether to print information of the selection process}

\item{Y}{Dataset y}

\item{n_predicors}{Amount of predictors to select}
}
\value{
A fitted model using the selected predictors.

A double value
}
\description{
Finds a subset of predictors using forward or backward selection
}
