% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_selection.R
\name{stepwise_selection}
\alias{stepwise_selection}
\title{Finds a subset of predictors using forward or backward selection}
\usage{
stepwise_selection(
  X,
  y,
  n_predictors,
  model = lm,
  direction = "forward",
  verbose = TRUE
)
}
\arguments{
\item{X}{Dataset X}

\item{y}{Dataset y}

\item{model}{Modeling function}

\item{direction}{Selection direction, either "forward" or "backward"}

\item{verbose}{Whether to print information of the selection process}

\item{n_predicors}{Amount of predictors to select}
}
\value{
A fitted model using the selected predictors.

A double value
}
\description{
Finds a subset of predictors using forward or backward selection
}
