% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_R2.R
\name{calculate_R2}
\alias{calculate_R2}
\title{Calculate model R-Squared}
\usage{
calculate_R2(y, yhat)
}
\arguments{
\item{y}{response vector}

\item{yhat}{fitted values}
}
\value{
R-Squared
}
\description{
Calculate model R-Squared
}
