% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_lasso.R
\name{coef.lasso}
\alias{coef.lasso}
\title{S3 plot function for lasso}
\usage{
\method{coef}{lasso}(lasso)
}
\arguments{
\item{lasso}{A lasso S3 Object}
}
\value{
plot
}
\description{
S3 plot function for lasso
}
