% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.ridge.R
\name{coef.ridge}
\alias{coef.ridge}
\title{Extract coefficients from an estimated ridge model}
\usage{
\method{coef}{ridge}(object, unscale = FALSE)
}
\arguments{
\item{object}{A ridge S3 Object}

\item{unscale}{If TRUE, the coefficients will be scaled back to the original
scale}
}
\value{
ridge coefficients from regobject
}
\description{
Extract coefficients from an estimated ridge model
}
