% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.stepwise_selection.R
\name{coef.stepwise_selection}
\alias{coef.stepwise_selection}
\title{Extracts the coefficients of a stepwise selection object.}
\usage{
\method{coef}{stepwise_selection}(object)
}
\arguments{
\item{object}{Stepwise selection object}
}
\value{
The coefficients.
}
\description{
Extracts the coefficients of a stepwise selection object.
}
