% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ridge.R
\name{plot.ridge}
\alias{plot.ridge}
\title{Plot the Ridge Regression Function over a predictor}
\usage{
\method{plot}{ridge}(ridgeobj, predictor)
}
\arguments{
\item{ridgeobj}{A ridge object}

\item{predictor}{The index or name of the predictor, which will be on the
x-axis}
}
\description{
Pass in a ridge object and select a predictor. The plot contains the original
points of the model and draw the estimation of y based solely on the selected
predictors. \cr
plot.ridge can plot real polynomials, so a model that was fitted for the formula
\code{y ~ x1 + x1^2 + 0.2*x1^5 + x2} could be plotted along x1.
}
\details{
See \code{vignette("plot", package = "regmodel")} for detailed worked examples.
}
