% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.stepwise_selection.R
\name{predict.stepwise_selection}
\alias{predict.stepwise_selection}
\title{Predict function for stepwise_selection()}
\usage{
\method{predict}{stepwise_selection}(selectionobject, newdata = NULL)
}
\description{
Predict function for stepwise_selection()
}
