% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmodel.R
\name{regmodel}
\alias{regmodel}
\title{Wrapper Function for the regmodelsuit package.}
\usage{
regmodel(formula = NULL, data = NULL, model = NULL, lambda = 0, cv = FALSE)
}
\arguments{
\item{data}{A data frame which contains the corresponding variables}

\item{model}{Specifies the model to be estimated.
c("ridge","lasso","forward","backward","LAR")}

\item{lambda}{penalty parameter for ridge and lasso estimation. If cv = TRUE
lambda will be ignored All of them are tested to find the optimal lambda
with regards to the chosen regmodel}

\item{formua}{A formula object which specifies the model}

\item{regmodel}{Logical value which specifies if cross validation for lambda should be used}
}
\value{
Model List
}
\description{
Wrapper Function for the regmodelsuit package.
}
