% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmodel.R
\name{regmodel}
\alias{regmodel}
\title{Wrapper Function for the regmodelsuit package.}
\usage{
regmodel(
  formula = NULL,
  data = NULL,
  model = NULL,
  lambda = NULL,
  cv = FALSE,
  ...
)
}
\arguments{
\item{formula}{A formula object which specifies the model}

\item{data}{A data frame which contains the corresponding variables}

\item{model}{Logical value which specifies if cross validation for lambda
should be used}

\item{lambda}{penalty parameter for ridge and lasso estimation. If cv = TRUE
lambda will be ignored all of them are tested to find the optimal lambda
with regard to the chosen regmodel}
}
\value{
Model List
}
\description{
Wrapper Function for the regmodelsuit package.
}
