% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmodelsuite-package.R
\docType{package}
\name{regmodelsuite-package}
\alias{regmodelsuite}
\alias{regmodelsuite-package}
\title{regmodelsuite: Estimate regression models}
\description{
Implements Ridge, Lasso, Forward selection, Backward selection and Least angle regression and also offers m-fold Cross Validation.
}
\author{
\strong{Maintainer}: 

Authors:
\itemize{
  \item Jonas Veit \email{mee322@uni-heidelberg.com}
  \item Pitt Junker \email{pitt.junker@stud.uni-heidelberg.com}
}

}
\keyword{internal}
