% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.R
\name{ridge}
\alias{ridge}
\title{Calculates the ridge regression}
\usage{
ridge(X, y, lambda = 0)
}
\arguments{
\item{X}{A matrix of predictor variables.}

\item{y}{A vector of response values.}

\item{lambda}{A numeric value for the regularization parameter.}
}
\value{
A vector of ridge regression coefficients.
}
\description{
This function calculates the coefficients of a ridge regression model
for a given predictor matrix \code{X} and response vector \code{y}.
}
