% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_cv.R
\name{ridge_cv}
\alias{ridge_cv}
\title{Calculates the Lambda for Ridge with m-fold Cross Validation}
\usage{
ridge_cv(X, y, m = 10, lambda)
}
\arguments{
\item{X}{Dataset X has the class "matrix".}

\item{y}{Target variable y has the class "matrix".}

\item{lambda}{Lambda is an atomic vector of double values.
cross_validation will find the optimal lambda among those. If no lambda is chosen,
then a default range of lambdas will be tested.}

\item{M}{Split X into M-many equally sized Groups.
m = 5 or m = 10 is recommended. The default is m = 10.}
}
\value{
A list with the mean squared prediction error, the optimal lambda and
the coefficients of the chosen regression model with the optimal lambda.
}
\description{
Calculates the Lambda for Ridge with m-fold Cross Validation
}
